api-gateway/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.25.1
python-dotenv==1.0.1
